Get Entity Classification

Retrieve entity classification from `http://classyfire.wishartlab.com/entities/'

get_classification(inchi_key)

Arguments

inchi_key
a character string of a valid InChIKey

Value

a tibble containing the following;

Level Classification level (kingdom, superclass, class and subclass)

Classification The compound classification

CHEMONT Chemical Ontology Identification code

Examples

# Valid InChI key where all four classification levels are available
get_classification('BRMWTNUJHUMWMS-LURJTMIESA-N')

#> ✔ BRMWTNUJHUMWMS-LURJTMIESA-N

#> # A tibble: 7 x 3
#>   Level      Classification                       CHEMONT          
#>   <chr>      <chr>                                <chr>            
#> 1 kingdom    Organic compounds                    CHEMONTID:0000000
#> 2 superclass Organic acids and derivatives        CHEMONTID:0000264
#> 3 class      Carboxylic acids and derivatives     CHEMONTID:0000265
#> 4 subclass   Amino acids, peptides, and analogues CHEMONTID:0000013
#> 5 level 5    Amino acids and derivatives          CHEMONTID:0000347
#> 6 level 6    Alpha amino acids and derivatives    CHEMONTID:0000060
#> 7 level 7    Histidine and derivatives            CHEMONTID:0004311

# Valid InChI key where only three classification levels are available
get_classification('MDHYEMXUFSJLGV-UHFFFAOYSA-N')

#> ✔ MDHYEMXUFSJLGV-UHFFFAOYSA-N

#> # A tibble: 3 x 3
#>   Level      Classification                      CHEMONT          
#>   <chr>      <chr>                               <chr>            
#> 1 kingdom    Organic compounds                   CHEMONTID:0000000
#> 2 superclass Benzenoids                          CHEMONTID:0002448
#> 3 class      Benzene and substituted derivatives CHEMONTID:0002279

# Invalid InChI key
get_classification('MDHYEMXUFSJLGV-UHFFFAOYSA-B')

#> ✖ MDHYEMXUFSJLGV-UHFFFAOYSA-B

# Using `dplyr` a vector of InChI Keys can be submitted and easily parsed
  library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following object is masked from ‘package:testthat’:
#> 
#>     matches

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

  library(purrr)

#> 
#> Attaching package: ‘purrr’

#> The following object is masked from ‘package:testthat’:
#> 
#>     is_null

#> The following object is masked from ‘package:magrittr’:
#> 
#>     set_names

  library(tidyr)

#> 
#> Attaching package: ‘tidyr’

#> The following object is masked from ‘package:magrittr’:
#> 
#>     extract

 keys <- c(
'BRMWTNUJHUMWMS-LURJTMIESA-N',
'XFNJVJPLKCPIBV-UHFFFAOYSA-N',
'TYEYBOSBBBHJIV-UHFFFAOYSA-N',
'AFENDNXGAFYKQO-UHFFFAOYSA-N',
'WHEUWNKSCXYKBU-QPWUGHHJSA-N',
'WHBMMWSBFZVSSR-GSVOUGTGSA-N')

 classification_list <- map(keys, get_classification)

#> ✔ BRMWTNUJHUMWMS-LURJTMIESA-N

#> ✔ XFNJVJPLKCPIBV-UHFFFAOYSA-N

#> ✔ TYEYBOSBBBHJIV-UHFFFAOYSA-N

#> ✔ AFENDNXGAFYKQO-UHFFFAOYSA-N

#> ✔ WHEUWNKSCXYKBU-QPWUGHHJSA-N

#> ✔ WHBMMWSBFZVSSR-GSVOUGTGSA-N

 classification_list <- map(classification_list, ~{select(.,-CHEMONT)})

 spread_tibble <- purrr:::map(classification_list, ~{
                  spread(., Level, Classification)
                  }) %>% bind_rows() %>% data.frame()

 rownames(spread_tibble) <- keys

 classification_df <-  data.frame(InChIKey = rownames(spread_tibble),
                               Kingdom = spread_tibble$kingdom,
                                SuperClass = spread_tibble$superclass,
                                Class = spread_tibble$class,
                                SubClass = spread_tibble$subclass)

 print(classification_df)

#>                      InChIKey           Kingdom                      SuperClass
#> 1 BRMWTNUJHUMWMS-LURJTMIESA-N Organic compounds   Organic acids and derivatives
#> 2 XFNJVJPLKCPIBV-UHFFFAOYSA-N Organic compounds      Organic nitrogen compounds
#> 3 TYEYBOSBBBHJIV-UHFFFAOYSA-N Organic compounds   Organic acids and derivatives
#> 4 AFENDNXGAFYKQO-UHFFFAOYSA-N Organic compounds   Organic acids and derivatives
#> 5 WHEUWNKSCXYKBU-QPWUGHHJSA-N Organic compounds Lipids and lipid-like molecules
#> 6 WHBMMWSBFZVSSR-GSVOUGTGSA-N Organic compounds   Organic acids and derivatives
#>                              Class                               SubClass
#> 1 Carboxylic acids and derivatives   Amino acids, peptides, and analogues
#> 2         Organonitrogen compounds                                 Amines
#> 3       Keto acids and derivatives Short-chain keto acids and derivatives
#> 4    Hydroxy acids and derivatives    Alpha hydroxy acids and derivatives
#> 5 Steroids and steroid derivatives                       Estrane steroids
#> 6    Hydroxy acids and derivatives     Beta hydroxy acids and derivatives